Windows: Factor out sqlite
(cherry picked from commit 3f6127b173949cb36557601a56bc15ae2c45a698)
