+++ pkg/graphdb/conn_sqlite3_linux.go
@@ -2,7 +2,11 @@
 
 package graphdb
 
-import "database/sql"
+import (
+	"database/sql"
+
+	_ "github.com/mattn/go-sqlite3" // registers sqlite
+)
 
 // NewSqliteConn opens a connection to a sqlite
 // database.
