--- pkg/graphdb/conn_sqlite3_unix.go
@@ -1,7 +1,0 @@
-// +build cgo,!windows
-
-package graphdb
-
-import (
-	_ "github.com/mattn/go-sqlite3" // registers sqlite
-)
