--- pkg/graphdb/conn_unsupported.go
@@ -1,8 +1,0 @@
-// +build !cgo
-
-package graphdb
-
-// NewSqliteConn return a new sqlite connection.
-func NewSqliteConn(root string) (*Database, error) {
-	panic("Not implemented")
-}
