+++ pkg/graphdb/unsupported.go
@@ -1,0 +1,3 @@
+// +build !cgo !linux
+
+package graphdb
